externalize orientation utils depending on board and pokemon entity
