+++ app/models/colyseus-models/player.ts
@@ -12,10 +12,10 @@ import {
   ConditionBasedEvolutionRule,
   carryOverPermanentStats
 } from "../../core/evolution-rules"
-import { FlowerPot, FlowerPots, MulchStockCaps } from "../../core/flower-pots"
+import { MulchStockCaps } from "../../core/flower-pots"
 import { PokemonEntity } from "../../core/pokemon-entity"
 import type GameState from "../../rooms/states/game-state"
-import { IPlayer, Role, Title } from "../../types"
+import { FlowerPot, FlowerPots, IPlayer, Role, Title } from "../../types"
 import { Ability } from "../../types/enum/Ability"
 import { DungeonPMDO } from "../../types/enum/Dungeon"
 import {
